// backend/src/types.ts
export interface CrashFilters {
  yearFrom?: number;
  yearTo?: number;
  region?: string;
  roadType?: string;
  speedEnvironment?: string;
  severity?: CrashSeverity;
}
// ...
export type CrashSeverity =
  | "Fatal Crash"
  | "Serious Crash"
  | "Minor Crash"
  | "Non-Injury Crash";